Use uint16.

+++ lib/debezium/decimal.go
@@ -8,9 +8,7 @@ import (
 )
 
 // EncodeDecimal is used to encode a string representation of a number to `org.apache.kafka.connect.data.Decimal`.
-func EncodeDecimal(value string, scale int) ([]byte, error) {
-	// TODO: Refactor scale to be uint16
-
+func EncodeDecimal(value string, scale uint16) ([]byte, error) {
 	bigFloatValue := new(big.Float)
 	if _, success := bigFloatValue.SetString(value); !success {
 		return nil, fmt.Errorf("unable to use %q as a floating-point number", value)
